refactor: colocate runner unit tests and retire remaining hand-rolled scaffolding
- Relocate the runner's unit tests into `diff`, `process`, and `report` beside the functions they pin
- Generalize the runner's `config_error` helper into `failed`, absorbing both `ParseError` reporting sites
- Delegate `testing::format_diagnostic` to the `Diagnostic::format` constructor it hand-rolled
- Reuse `format_diagnostic` in cache tests, narrowing `entry` to the `&str` every caller passes
- Drop the pipeline `SentinelRule`'s never-applied `edits` field
- Adopt `testing::range` over hand-built ranges in the aligner's emission tests
- Pin `format -` writing the formatted rewrite to stdout when stdin needs changes
- Correct the `strip-trailing-commas` caption, which claimed the rule strips only single-line collections
