+++ site/rules/strip-trailing-commas.md
@@ -1,7 +1,7 @@
 ---
 category : auto-fix
 family   : formatting
-caption  : "removes trailing commas from inside single-line collections."
+caption  : "removes trailing commas from collections, signatures, calls, and every other bracketed container."
 related  : [collection-layout, align-colons]
 layout   : doc
 ---
