+++ src/cache/mod.rs
@@ -20,14 +20,13 @@ pub use records::{CacheEntry, CacheInfo, CleanReport, Rewrite};
 
 #[cfg(test)]
 mod tests {
-    use ruff_diagnostics::{Edit, Fix};
     use tempfile::TempDir;
 
     use super::key::CACHE_FORMAT_VERSION;
     use super::*;
-    use crate::diagnostics::{Diagnostic, Severity};
+    use crate::diagnostics::Diagnostic;
     use crate::rule::RuleId;
-    use crate::testing::range;
+    use crate::testing::{format_diagnostic, range};
 
     const CONFIG_A: &str = "code-line-length = 88\n";
     const CONFIG_B: &str = "code-line-length = 100\n";
@@ -41,25 +40,16 @@ mod tests {
         }
     }
 
-    fn entry(formatted: Option<&str>) -> CacheEntry {
+    fn entry(formatted: &str) -> CacheEntry {
         CacheEntry {
             // The rule must be a registered slug, because `RuleId`
             // deserializes through the registry and an unknown slug
             // fails the entry's round-trip.
             diagnostics: vec![Diagnostic {
-                fix: Some(Fix::safe_edit(Edit::range_replacement(
-                    "y".into(),
-                    range(0, 1),
-                ))),
-                message: "rewrite".into(),
-                range: range(0, 1),
                 rule: RuleId::from("align-equals"),
-                severity: Severity::Format,
+                ..format_diagnostic(range(0, 1))
             }],
-            rewrite: match formatted {
-                Some(text) => Rewrite::Changed(text.to_owned()),
-                None => Rewrite::Unchanged,
-            },
+            rewrite: Rewrite::Changed(formatted.to_owned()),
         }
     }
 
@@ -121,7 +111,7 @@ mod tests {
         let tmp = TempDir::new().expect("tempdir");
         let cache = cache_in(&tmp, 100);
         let key = CacheKey::compute(b"x = 1\n", CONFIG_A);
-        cache.insert(&key, &entry(Some("y = 1\n")));
+        cache.insert(&key, &entry("y = 1\n"));
         let report = cache.clean().expect("cleans");
         assert_eq!(report.entries, 1);
         assert!(report.bytes > 0);
@@ -143,9 +133,9 @@ mod tests {
         let cache = cache_in(&tmp, 100);
         let key_old = CacheKey::compute(b"x = 1\n", CONFIG_A);
         let key_new = CacheKey::compute(b"y = 2\n", CONFIG_A);
-        cache.insert(&key_old, &entry(Some("a = 1\n")));
+        cache.insert(&key_old, &entry("a = 1\n"));
         std::thread::sleep(std::time::Duration::from_millis(20));
-        cache.insert(&key_new, &entry(Some("b = 2\n")));
+        cache.insert(&key_new, &entry("b = 2\n"));
 
         let tightened = Cache {
             max_size_bytes: 0,
@@ -162,7 +152,7 @@ mod tests {
         let tmp = TempDir::new().expect("tempdir");
         let cache = cache_in(&tmp, 100);
         let key = CacheKey::compute(b"x = 1\n", CONFIG_A);
-        cache.insert(&key, &entry(Some("y = 1\n")));
+        cache.insert(&key, &entry("y = 1\n"));
         let report = cache.compact();
         assert_eq!(report.entries, 0);
         assert_eq!(report.bytes, 0);
@@ -172,14 +162,8 @@ mod tests {
     fn info_counts_entries_and_byte_total() {
         let tmp = TempDir::new().expect("tempdir");
         let cache = cache_in(&tmp, 100);
-        cache.insert(
-            &CacheKey::compute(b"x = 1\n", CONFIG_A),
-            &entry(Some("y = 1\n")),
-        );
-        cache.insert(
-            &CacheKey::compute(b"x = 2\n", CONFIG_A),
-            &entry(Some("y = 2\n")),
-        );
+        cache.insert(&CacheKey::compute(b"x = 1\n", CONFIG_A), &entry("y = 1\n"));
+        cache.insert(&CacheKey::compute(b"x = 2\n", CONFIG_A), &entry("y = 2\n"));
         let info = cache.info();
         assert_eq!(info.entries, 2);
         assert!(info.bytes > 0);
@@ -214,9 +198,9 @@ mod tests {
         let cache = cache_in(&tmp, 0);
         let key_old = CacheKey::compute(b"x = 1\n", CONFIG_A);
         let key_new = CacheKey::compute(b"y = 2\n", CONFIG_A);
-        cache.insert(&key_old, &entry(Some("a = 1\n")));
+        cache.insert(&key_old, &entry("a = 1\n"));
         std::thread::sleep(std::time::Duration::from_millis(20));
-        cache.insert(&key_new, &entry(Some("b = 2\n")));
+        cache.insert(&key_new, &entry("b = 2\n"));
         assert!(cache.lookup(&key_old).is_none());
     }
 
@@ -225,7 +209,7 @@ mod tests {
         let tmp = TempDir::new().expect("tempdir");
         let cache = cache_in(&tmp, 100);
         let key = CacheKey::compute(b"x = 1\n", CONFIG_A);
-        cache.insert(&key, &entry(Some("y = 1\n")));
+        cache.insert(&key, &entry("y = 1\n"));
         let tmp_count = fs_err::read_dir(&cache.root)
             .expect("read_dir")
             .flatten()
@@ -252,7 +236,7 @@ mod tests {
         let tmp = TempDir::new().expect("tempdir");
         let cache = cache_in(&tmp, 100);
         let key = CacheKey::compute(b"x = 1\n", CONFIG_A);
-        let original = entry(Some("y = 1\n"));
+        let original = entry("y = 1\n");
         cache.insert(&key, &original);
         let recovered = cache.lookup(&key).expect("hit");
         assert_eq!(recovered, original);
