+++ src/cli/runner/report.rs
@@ -185,3 +185,277 @@ pub(super) fn summarize(
         resolved => Some(resolved),
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use assert_matches::assert_matches;
+    use rstest::rstest;
+    use ruff_diagnostics::{Edit, Fix};
+    use ruff_text_size::TextRange;
+
+    use super::*;
+    use crate::diagnostics::Severity;
+    use crate::rule::RuleId;
+    use crate::source::Source;
+    use crate::testing::{parse, range};
+
+    struct FailingWriter;
+
+    impl Write for FailingWriter {
+        fn flush(&mut self) -> io::Result<()> {
+            Ok(())
+        }
+
+        fn write(&mut self, _: &[u8]) -> io::Result<usize> {
+            Err(io::Error::other("simulated write failure"))
+        }
+    }
+
+    fn diagnostic(severity: Severity, range: TextRange, slug: &'static str) -> Diagnostic {
+        Diagnostic {
+            fix: matches!(severity, Severity::Format)
+                .then(|| Fix::safe_edit(Edit::range_replacement("y".into(), range))),
+            message: "test".into(),
+            range,
+            rule: RuleId::from(slug),
+            severity,
+        }
+    }
+
+    fn outcome_with(source: Source, diagnostics: Vec<Diagnostic>) -> FileOutcome {
+        FileOutcome::Done {
+            cached: false,
+            diagnostics,
+            file: source.source_file().clone(),
+            rewrite: Rewrite::Skipped,
+        }
+    }
+
+    #[test]
+    fn check_outcomes_with_failed_parse_takes_higher_status() {
+        let source = parse("x = 1\n");
+        let range = range(0, 1);
+        let outcomes = vec![
+            outcome_with(
+                source,
+                vec![diagnostic(Severity::Format, range, "synthetic-format")],
+            ),
+            FileOutcome::Failed(ExitStatus::ParseError),
+        ];
+
+        let status = status_from_outcomes(&outcomes, false);
+
+        assert_eq!(status, ExitStatus::ParseError);
+    }
+
+    #[test]
+    fn check_outcomes_with_lint_and_format_returns_lint_violation() {
+        let source = parse("x = 1\n");
+        let range = range(0, 1);
+        let diagnostics = vec![
+            diagnostic(Severity::Format, range, "synthetic-format"),
+            diagnostic(Severity::Lint, range, "synthetic-lint"),
+        ];
+        let outcomes = vec![outcome_with(source, diagnostics)];
+
+        let status = status_from_outcomes(&outcomes, false);
+
+        assert_eq!(status, ExitStatus::LintViolation);
+    }
+
+    #[test]
+    fn check_outcomes_with_synthetic_lint_returns_lint_violation() {
+        let source = parse("x = 1\n");
+        let diagnostics = vec![diagnostic(Severity::Lint, range(0, 1), "synthetic-lint")];
+        let outcomes = vec![outcome_with(source, diagnostics)];
+
+        let status = status_from_outcomes(&outcomes, false);
+
+        assert_eq!(status, ExitStatus::LintViolation);
+    }
+
+    #[test]
+    fn emit_outcomes_propagates_writer_failure() {
+        let source = parse("x = 1\n");
+        let diags = vec![diagnostic(
+            Severity::Format,
+            range(0, 1),
+            "synthetic-format",
+        )];
+        let outcomes = vec![outcome_with(source, diags)];
+        let result = emit_outcomes(
+            &outcomes,
+            OutputFormat::Json,
+            &mut FailingWriter,
+            &EmitterSummary::default(),
+        );
+        assert!(result.is_err());
+    }
+
+    #[rstest]
+    fn emit_outcomes_renders_each_output_format(
+        #[values(
+            OutputFormat::Github,
+            OutputFormat::Json,
+            OutputFormat::Sarif,
+            OutputFormat::Text
+        )]
+        format: OutputFormat,
+    ) {
+        let source = parse("x = 1\n");
+        let outcomes = vec![outcome_with(source, Vec::new())];
+        let mut buf = Vec::new();
+        emit_outcomes(&outcomes, format, &mut buf, &EmitterSummary::default()).expect("emits");
+    }
+
+    #[test]
+    fn emitter_summary_counts_visited_changed_diagnostics_and_rules() {
+        let range = range(0, 1);
+        let mut changed = outcome_with(
+            parse("x = 1\n"),
+            vec![
+                diagnostic(Severity::Format, range, "align-equals"),
+                diagnostic(Severity::Lint, range, "reassigned-constants"),
+            ],
+        );
+        if let FileOutcome::Done { rewrite, .. } = &mut changed {
+            *rewrite = Rewrite::Changed("x   = 1\n".to_owned());
+        }
+        let clean = outcome_with(parse("y = 2\n"), Vec::new());
+        let outcomes = vec![changed, clean, FileOutcome::Failed(ExitStatus::ParseError)];
+
+        let summary = emitter_summary(&outcomes);
+
+        assert_eq!(summary.files_visited, 2);
+        assert_eq!(summary.files_changed, 1);
+        assert_eq!(summary.files_with_diagnostics, 1);
+        assert_eq!(summary.diagnostics_total, 2);
+        assert_eq!(summary.rules_fired[&RuleId::from("align-equals")], 1);
+        assert_eq!(
+            summary.rules_fired[&RuleId::from("reassigned-constants")],
+            1
+        );
+    }
+
+    #[test]
+    fn emitter_summary_tallies_repeated_rule_occurrences() {
+        let range = range(0, 1);
+        let outcome = outcome_with(
+            parse("x = 1\n"),
+            vec![
+                diagnostic(Severity::Format, range, "align-equals"),
+                diagnostic(Severity::Format, range, "align-equals"),
+            ],
+        );
+
+        let summary = emitter_summary(std::slice::from_ref(&outcome));
+
+        assert_eq!(summary.rules_fired[&RuleId::from("align-equals")], 2);
+    }
+
+    #[test]
+    fn file_changed_counts_a_changed_rewrite_or_a_skipped_format_diagnostic() {
+        let range = range(0, 1);
+        let format = vec![diagnostic(Severity::Format, range, "synthetic-format")];
+        let lint = vec![diagnostic(Severity::Lint, range, "synthetic-lint")];
+
+        assert!(file_changed(&[], &Rewrite::Changed("x = 1\n".to_owned())));
+        assert!(file_changed(&format, &Rewrite::Skipped));
+        assert!(!file_changed(&format, &Rewrite::Unchanged));
+        assert!(!file_changed(&lint, &Rewrite::Skipped));
+        assert!(!file_changed(&[], &Rewrite::Skipped));
+    }
+
+    #[test]
+    fn report_verbose_prints_bypassed_when_cache_disabled() {
+        let mut buf = Vec::new();
+        report_verbose(&[], false, &mut buf);
+        assert_eq!(String::from_utf8(buf).expect("utf-8"), "cache: bypassed\n");
+    }
+
+    #[test]
+    fn report_verbose_prints_hit_and_miss_counts() {
+        let make = |cached: bool| {
+            let source = parse("x = 1\n");
+            let mut o = outcome_with(source, Vec::new());
+            if let FileOutcome::Done { cached: c, .. } = &mut o {
+                *c = cached;
+            }
+            o
+        };
+        let outcomes = vec![
+            make(true),
+            make(true),
+            make(false),
+            FileOutcome::Failed(ExitStatus::Clean),
+        ];
+        let mut buf = Vec::new();
+        report_verbose(&outcomes, true, &mut buf);
+        assert_eq!(
+            String::from_utf8(buf).expect("utf-8"),
+            "cache: 2 hits, 1 misses, 3 files\n",
+        );
+    }
+
+    #[test]
+    fn status_from_outcomes_clears_format_change_for_an_unchanged_rewrite() {
+        let source = parse("x = 1\n");
+        let range = range(0, 1);
+        let mut outcome = outcome_with(
+            source,
+            vec![
+                diagnostic(Severity::Format, range, "synthetic-format"),
+                diagnostic(Severity::Lint, range, "synthetic-lint"),
+            ],
+        );
+        if let FileOutcome::Done { rewrite, .. } = &mut outcome {
+            *rewrite = Rewrite::Unchanged;
+        }
+        let outcomes = vec![outcome];
+
+        assert_eq!(
+            status_from_outcomes(&outcomes, false),
+            ExitStatus::LintViolation,
+        );
+    }
+
+    #[test]
+    fn status_from_outcomes_demotes_format_change_when_demoted() {
+        let source = parse("x = 1\n");
+        let outcomes = vec![outcome_with(
+            source,
+            vec![diagnostic(
+                Severity::Format,
+                range(0, 1),
+                "synthetic-format",
+            )],
+        )];
+        assert_eq!(status_from_outcomes(&outcomes, true), ExitStatus::Clean);
+        assert_eq!(
+            status_from_outcomes(&outcomes, false),
+            ExitStatus::FormatChange,
+        );
+    }
+
+    #[test]
+    fn summarize_reports_diagnostics_alongside_a_failure() {
+        let source = parse("x = 1\n");
+        let outcomes = vec![
+            outcome_with(
+                source,
+                vec![diagnostic(Severity::Lint, range(0, 1), "synthetic-lint")],
+            ),
+            FileOutcome::Failed(ExitStatus::ParseError),
+        ];
+        assert_matches!(
+            summarize(&outcomes, &emitter_summary(&outcomes), Mode::Check),
+            Some(Summary::Diagnostics { files: 1, total: 1 })
+        );
+    }
+
+    #[test]
+    fn summarize_suppresses_clean_summary_when_a_file_failed() {
+        let outcomes = vec![FileOutcome::Failed(ExitStatus::ParseError)];
+        assert!(summarize(&outcomes, &emitter_summary(&outcomes), Mode::Check).is_none());
+    }
+}
